# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a API documentation of BE Revend",
        "title": "BE Revend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Rizal Arfiyan",
            "url": "https://rizalrfiyan.com",
            "email": "rizal.arfiyan.23@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "description": "Base Home",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "home"
                ],
                "summary": "Get Base Home based on parameter",
                "operationId": "get-base-home",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/auth/google": {
            "get": {
                "description": "Auth Google Redirection",
                "tags": [
                    "auth"
                ],
                "summary": "Get Auth Google Redirection based on parameter",
                "operationId": "get-auth-google",
                "responses": {
                    "307": {
                        "description": "Temporary Redirect"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/auth/google/callback": {
            "get": {
                "description": "Auth Google Callback Redirection",
                "tags": [
                    "auth"
                ],
                "summary": "Get Auth Google Callback Redirection based on parameter",
                "operationId": "get-auth-google-callback",
                "responses": {
                    "307": {
                        "description": "Temporary Redirect"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "AccessToken": []
                    }
                ],
                "description": "Auth Me",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Get Auth Me based on parameter",
                "operationId": "get-auth-me",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.AuthToken"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/auth/otp": {
            "post": {
                "description": "Auth Send OTP",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Post Auth Send OTP based on parameter",
                "operationId": "post-auth-send-otp",
                "parameters": [
                    {
                        "description": "Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AuthSendOTP"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/auth/otp/verification": {
            "post": {
                "description": "Auth Verification",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Post Auth Verification based on parameter",
                "operationId": "post-auth-otp-verification",
                "parameters": [
                    {
                        "description": "Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AuthOTPVerification"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Auth Register",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Post Auth Register based on parameter",
                "operationId": "post-auth-register",
                "parameters": [
                    {
                        "description": "Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AuthRegister"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/auth/verification": {
            "post": {
                "description": "Auth Verification",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Post Auth Verification based on parameter",
                "operationId": "post-auth-verification",
                "parameters": [
                    {
                        "description": "Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AuthVerification"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.AuthVerification"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "constants.AuthVerificationStep": {
            "type": "integer",
            "enum": [
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "AuthVerificationRegister",
                "AuthVerificationOtp",
                "AuthVerificationDone"
            ]
        },
        "models.AuthToken": {
            "type": "object",
            "properties": {
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/sql.Role"
                }
            }
        },
        "request.AuthOTPVerification": {
            "type": "object",
            "properties": {
                "otp": {
                    "type": "string",
                    "example": "651721"
                },
                "token": {
                    "type": "string",
                    "example": "2YbPyusF2G06BFQLamoKFXvGgPd"
                }
            }
        },
        "request.AuthRegister": {
            "type": "object",
            "required": [
                "first_name",
                "token"
            ],
            "properties": {
                "first_name": {
                    "type": "string",
                    "minLength": 3,
                    "example": "Rizal"
                },
                "last_name": {
                    "type": "string",
                    "minLength": 3,
                    "example": "Arfiyan"
                },
                "phone_number": {
                    "type": "string",
                    "example": "62895377233002"
                },
                "token": {
                    "type": "string",
                    "example": "2YbPyusF2G06BFQLamoKFXvGgPd"
                }
            }
        },
        "request.AuthSendOTP": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "phone_number": {
                    "type": "string",
                    "example": "62895377233002"
                },
                "token": {
                    "type": "string",
                    "example": "2YbPyusF2G06BFQLamoKFXvGgPd"
                }
            }
        },
        "request.AuthVerification": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string",
                    "example": "2YbPyusF2G06BFQLamoKFXvGgPd"
                }
            }
        },
...
        "response.AuthVerification": {
            "type": "object",
            "properties": {
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "remaining_time": {
                    "type": "integer"
                },
                "step": {
                    "$ref": "#/definitions/constants.AuthVerificationStep"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "response.BaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 999
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Message!"
                }
            }
        },
...
        "sql.Role": {
            "type": "string",
            "enum": [
                "admin",
                "guest"
            ],
            "x-enum-varnames": [
                "RoleAdmin",
                "RoleGuest"
            ]
        }
    },
    "securityDefinitions": {
        "AccessToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}