# internal/sql/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.23.0

package sql

import (
	"database/sql/driver"
	"fmt"

	"github.com/jackc/pgx/v5/pgtype"
)

type Role string

const (
	RoleAdmin Role = "admin"
	RoleGuest Role = "guest"
)

func (e *Role) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = Role(s)
	case string:
		*e = Role(s)
	default:
		return fmt.Errorf("unsupported scan type for Role: %T", src)
	}
	return nil
}

type NullRole struct {
	Role  Role
	Valid bool // Valid is true if Role is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullRole) Scan(value interface{}) error {
	if value == nil {
		ns.Role, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.Role.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullRole) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.Role), nil
}

type Device struct {
	ID        pgtype.UUID
	Token     string
	Name      string
	Location  string
	CreatedAt pgtype.Timestamp
	UpdatedAt pgtype.Timestamp
	DeletedBy pgtype.UUID
	DeletedAt pgtype.Timestamp
}

type History struct {
	ID        pgtype.UUID
	UserID    pgtype.UUID
	DeviceID  pgtype.UUID
	Success   int32
	Failed    int32
	CreatedAt pgtype.Timestamp
}

type User struct {
	ID          pgtype.UUID
	FirstName   string
	LastName    pgtype.Text
	PhoneNumber string
	GoogleID    pgtype.Text
	Identity    string
	Role        Role
	CreatedAt   pgtype.Timestamp
	UpdatedAt   pgtype.Timestamp
	DeletedBy   pgtype.UUID
	DeletedAt   pgtype.Timestamp
}